--- src/main/java/site/metacoding/miniproject2/dto/CareersCodeRespDto.java
@@ -1,8 +1,0 @@
-package site.metacoding.miniproject2.dto;
-
-@Setter
-@Getter
-public class CareersCodeRespDto {
-	private Integer id;
-	private String name;
-}
