--- src/main/java/site/metacoding/miniproject2/dto/RegionsCodeRespDto.java
@@ -1,8 +1,0 @@
-package site.metacoding.miniproject2.dto;
-
-@Setter
-@Getter
-public class RegionsCodeRespDto {
-	private Integer id;
-	private String name;
-}
