--- src/main/java/site/metacoding/miniproject2/dto/SkillsCodeRespDto.java
@@ -1,8 +1,0 @@
-package site.metacoding.miniproject2.dto;
-
-@Setter
-@Getter
-public class SkillsCodeRespDto {
-	private Integer id;
-	private String name;
-}
